+++ racing_data/entity.py
@@ -6,11 +6,12 @@ import pytz
 class Entity(dict):
     """Common functionality for racing entities"""
     
-    def __init__(self, provider, *args, **kwargs):
+    def __init__(self, provider, property_cache, *args, **kwargs):
 
         super(Entity, self).__init__(*args, **kwargs)
 
         self.provider = provider
+        self.property_cache = property_cache
 
         if not 'created_at' in self:
             self['created_at'] = self['updated_at'] = datetime.now(pytz.utc)
@@ -27,6 +28,17 @@ class Entity(dict):
         """Expire entities sourced from an incompatible scraper version"""
 
         return not self.provider.scraper.is_compatible_with(self['scraper_version'])
+
+    def get_cached_property(self, key, source_method, *source_args, **source_kwargs):
+        """Get a cached property value, or source, cache and return it if necessary"""
+
+        if self.property_cache is None:
+            self.property_cache = {}
+
+        if key not in self.property_cache:
+            self.property_cache = source_method(*source_args, **source_kwargs)
+
+        return self.property_cache[key]
 
     def is_equivalent_to(self, other_entity):
         """This entity is equivalent to other_entity if both are equal"""