+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Get meets by date (from @justjasongreen)
 - Get races by meet (from @justjasongreen)
+- Get runners by race (from @justjasongreen)
 
 ## [1.0.0a0] - 2016-07-21
 ### Added